feat(api): decouple non-route methods
